requirements.txt
absl-py==0.9.0
actionlib==1.12.1
aiohttp==3.6.2
alabaster==0.7.12
angles==1.9.12
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.12.0
argon2-cffi==20.1.0
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
autopep8==1.5.3
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
bondpy==1.8.5
box2d-py==2.3.8
breathe==4.19.2
Brlapi==0.6.6
bs4==0.0.1
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.28
catkin-pkg==0.4.22
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.3.0
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
controller-manager==0.18.2
controller-manager-msgs==0.18.2
cryptography==2.9.2
cupshelpers==1.0
cv-bridge==1.13.0
cycler==0.10.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
dill==0.3.1.1
distlib==0.3.1
distro==1.5.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.16
doxypypy==0.8.8.6
dynamic-reconfigure==1.6.3
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gcovr==4.2
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.14
gitdb==4.0.5
GitPython==3.1.7
gpytorch==1.1.1
gym==0.17.2
h5py==2.10.0
httplib2==0.9.2
idna==2.9
idna-ssl==1.1.0
image-geometry==1.13.0
imagesize==1.2.0
importlib-metadata==1.7.0
importlib-resources==3.0.0
interactive-markers==1.11.5
ipdb==0.13.3
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jax==0.1.75
jaxlib @ https://storage.googleapis.com/jax-releases/cuda102/jaxlib-0.1.52-cp36-none-manylinux2010_x86_64.whl
jedi==0.17.0
Jinja2==2.11.2
joblib==0.16.0
joint-state-publisher==1.12.15
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-server-proxy==1.5.0
kdl-parser-py==1.13.1
Keras==2.4.3
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
language-selector==0.1
laser-geometry==1.6.5
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lightargs==1.23
louis==3.5.0
lxml==4.5.2
m2r==0.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.2.1
message-filters==1.14.9
mistune==0.8.4
mock==4.0.2
moveit-commander==1.0.6
moveit-ros-planning-interface==1.0.6
moveit-ros-visualization==1.0.6
mpmath==1.1.0
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.7
ndg-httpsclient==0.5.1
netifaces==0.10.4
notebook==6.0.3
numpy==1.19.1
oauth==1.0.1
olefile==0.45.1
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.0.0
ptyprocess==0.6.0
pyasn1==0.4.8
pybullet==2.7.9
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pyglet==1.5.0
Pygments==2.6.1
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.16.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
python-qt-binding==0.4.3
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==4.7.4
QtPy==1.9.0
recommonmark==0.6.0
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
resource-retriever==1.12.6
roboball2d==0.1.6
rope==0.17.0
rosbag==1.14.9
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosdep==0.19.0
rosdistro==0.8.2
rosgraph==1.14.9
roslaunch==1.14.9
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.9
rosmake==1.14.9
rosmaster==1.14.9
rosmsg==1.14.9
rosnode==1.14.9
rosparam==1.14.9
rospkg==1.2.8
rospy==1.14.9
rosservice==1.14.9
rostest==1.14.9
rostopic==1.14.9
rosunit==1.14.9
roswtf==1.14.9
rqt-action==0.4.9
rqt-bag==0.4.14
rqt-bag-plugins==0.4.14
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.9
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.13.13
schedule==0.6.0
scikit-learn==0.23.2
scipy==1.5.0
scour==0.36
screen-resolution-extra==0.0.0
seaborn==0.10.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sensor-msgs==1.12.7
simpervisor==0.3
simplejson==3.13.2
six==1.15.0
sklearn==0.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
smmap==3.0.4
snowballstemmer==2.0.0
sound-play==0.3.7
soupsieve==2.0.1
Sphinx==3.2.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-moderncmakedomain==3.13
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
srdfdom==0.5.2
ssh-import-id==5.7
sympy==1.6
system-service==0.3
systemd-python==234
terminado==0.8.3
testpath==0.4.4
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
threadpoolctl==2.1.0
toml==0.10.1
topic-tools==1.14.9
torch==1.5.1
torchvision==0.6.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
treep==1.47
txaio==20.4.1
typing-extensions==3.7.4.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urdfdom-py==0.4.3
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.0.30
vpython==7.6.1
wadllib==1.3.2
wcwidth==0.2.4
webencodings==0.5.1
widgetsnbextension==3.5.1
xacro==1.13.8
xkit==0.0.0
yarl==1.4.2
zipp==3.1.0
zope.interface==4.3.2
